GoalSmartSuggestions removed useGoals refactored GoalCard updated amount issue
